backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2.0,<3
pymysql>=1.1,<2
pydantic>=2,<3
pydantic-settings>=2,<3
email-validator>=2,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
bcrypt>=4,<5
python-multipart>=0.0.9,<1
PyPDF2>=3.0,<4
python-docx>=1.1,<2
